+++ internal/config/config.go
@@ -164,6 +164,7 @@ type Options struct {
 	MinifyIdentifiers bool
 	MangleSyntax      bool
 	CodeSplitting     bool
+	WatchMode         bool
 
 	// Setting this to true disables warnings about code that is very likely to
 	// be a bug. This is used to ignore issues inside "node_modules" directories.